requirements.txt
# Working requirements for Stock Prediction App
# Tested on Python 3.12.9 - January 2025

# Web Framework
streamlit==1.39.0

# Data Processing
pandas==2.3.2
numpy==2.2.6
yfinance==0.2.66

# ML Models
xgboost==3.0.5
lightgbm==4.6.0
scikit-learn==1.7.2
scipy==1.16.2

# Technical Indicators
ta-lib==0.6.7
pandas-ta==0.4.71b0
numba==0.61.2

# Visualization
plotly==6.3.0

# Database
sqlalchemy==2.0.43
redis==6.4.0
psycopg2-binary==2.9.10

# Async
aiohttp==3.12.15

# Utils
python-dotenv==1.1.1
pyyaml==6.0.2
tqdm==4.67.1
joblib==1.5.2

# OpenAI for news sentiment
openai==1.108.0

# Testing
pytest==8.4.2

# Optional: Deep Learning (add if needed)
# torch>=2.0.0
# transformers>=4.35.0
# accelerate>=0.25.0
# chronos-torch>=1.1.0
